Making a copy of the config object to not modify it.
Motivation:
Currently, the createLogger function will modify the passed in config
object. This might not be expected by the caller and could also cause
issue if this instance is later updated from the outide.
Instead a copy of the config should be taken and used by the logger.

Modifications:
Added a copy of when the config object is not a string.

Result:
The passed-in config object will not be modified by the createLogger
function.